+++ packages/article-image/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.16.7"></a>
+## [0.16.7](https://github.com/newsuk/times-components/compare/@times-components/article-image@0.16.6...@times-components/article-image@0.16.7) (2018-05-29)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-image
+
 <a name="0.16.6"></a>
 ## [0.16.6](https://github.com/newsuk/times-components/compare/@times-components/article-image@0.16.5...@times-components/article-image@0.16.6) (2018-05-25)