+++ packages/article-summary/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.24.7"></a>
+## [0.24.7](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.24.6...@times-components/article-summary@0.24.7) (2018-05-29)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-summary
+
 <a name="0.24.6"></a>
 ## [0.24.6](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.24.5...@times-components/article-summary@0.24.6) (2018-05-25)