+++ packages/gestures/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.4.6"></a>
+## [1.4.6](https://github.com/newsuk/times-components/compare/@times-components/gestures@1.4.5...@times-components/gestures@1.4.6) (2018-05-29)
+
+
+
+
+**Note:** Version bump only for package @times-components/gestures
+
 <a name="1.4.5"></a>
 ## [1.4.5](https://github.com/newsuk/times-components/compare/@times-components/gestures@1.4.4...@times-components/gestures@1.4.5) (2018-05-25)