+++ packages/gradient/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.12.1"></a>
+## [0.12.1](https://github.com/newsuk/times-components/compare/@times-components/gradient@0.12.0...@times-components/gradient@0.12.1) (2018-05-29)
+
+
+
+
+**Note:** Version bump only for package @times-components/gradient
+
 <a name="0.12.0"></a>
 # [0.12.0](https://github.com/newsuk/times-components/compare/@times-components/gradient@0.11.0...@times-components/gradient@0.12.0) (2018-05-25)