+++ packages/icons/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.8.6"></a>
+## [0.8.6](https://github.com/newsuk/times-components/compare/@times-components/icons@0.8.5...@times-components/icons@0.8.6) (2018-05-29)
+
+
+
+
+**Note:** Version bump only for package @times-components/icons
+
 <a name="0.8.5"></a>
 ## [0.8.5](https://github.com/newsuk/times-components/compare/@times-components/icons@0.8.4...@times-components/icons@0.8.5) (2018-05-25)