+++ packages/pull-quote/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.7.6"></a>
+## [0.7.6](https://github.com/newsuk/times-components/compare/@times-components/pull-quote@0.7.5...@times-components/pull-quote@0.7.6) (2018-05-29)
+
+
+
+
+**Note:** Version bump only for package @times-components/pull-quote
+
 <a name="0.7.5"></a>
 ## [0.7.5](https://github.com/newsuk/times-components/compare/@times-components/pull-quote@0.7.4...@times-components/pull-quote@0.7.5) (2018-05-25)