+++ packages/styleguide/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.8.2"></a>
+## [0.8.2](https://github.com/newsuk/times-components/compare/@times-components/styleguide@0.8.1...@times-components/styleguide@0.8.2) (2018-05-29)
+
+
+
+
+**Note:** Version bump only for package @times-components/styleguide
+
 <a name="0.8.1"></a>
 ## [0.8.1](https://github.com/newsuk/times-components/compare/@times-components/styleguide@0.8.0...@times-components/styleguide@0.8.1) (2018-05-25)